add Ubuntu 20.04 + MuJoCo 3.6.0 support
- CMakeLists: cross-platform build with USE_DDS option (default OFF on Linux),
  auto-detect MuJoCo at ~/.mujoco/mujoco-3.6.0, manual OpenCV paths to avoid
  libgphoto2 conflict, link pthread and OpenGL on Linux
- simulate/CMakeLists: exclude macOS .mm files on Linux
- simulate/simulate.cc: compat no-op for removed MJDATA_POINTERS_PREAMBLE (MuJoCo 3.4+)
- planner: replace Eigen::Vector<T,N> with Matrix<T,N,1> for Eigen 3.3 compat,
  fix brace-init macros to use comma-initializer, add missing <mutex> include
- src/main.cpp: guard all DDS code with #ifdef USE_DDS, add LoadMuJoCoPlugins()
  to auto-load OBJ/STL decoders from MuJoCo installation
- src/RobotSim.cpp/hpp: fix mjtSize vs int type mismatch (MuJoCo 3.6 API change),
  replace opencv.hpp with specific component headers
- tools/CMakeLists: skip dds_control_client when USE_DDS=OFF

+++ src/RobotSim.cpp
@@ -188,7 +188,7 @@ void RobotSim::applyControlVector(const std::vector<double>& control) {
   std::lock_guard<std::mutex> lock(sim_mutex);
   if (!m || !d)
     return;
-  const int count = std::min(static_cast<int>(control.size()), m->nu);
+  const int count = std::min(static_cast<int>(control.size()), static_cast<int>(m->nu));
   for (int i = 0; i < count; ++i)
     d->ctrl[i] = control[i];
 }
